Fix the beamwidth factor in antenna string
The "beamwidth" is not in degrees but rather as a factor of the usual
lambda / D width. It therefore ranges from about 1.03 for uniform
"hard" illumination to about 1.22 for Gaussian "soft" illumination.
Given that this is a communication dish and not knowing more about the
feed I would guess 1.03 for starters.

This ensures that beam fitting gets a proper expected beamwidth.